Add event to world and made it rollbackable
